+++ spb/backends/bokeh/bokeh.py
@@ -86,12 +86,13 @@ class BokehBackend(Plot):
         NyquistLineSeries: NyquistRenderer,
     }
 
-    pole_line_kw = {"line_color": "#000000", "line_dash": "dotted"}
-    grid_line_kw = {"line_color": "#aaa", "line_dash": "dotted"}
-    sgrid_line_kw = {"line_color": "#aaa", "line_dash": "dotted"}
-    ngrid_line_kw = {"line_color": "#aaa", "line_dash": "dotted"}
-    mcircles_line_kw = {"line_color": "#aaa", "line_dash": "dotted"}
-
+    pole_line_kw = param.Dict(default={
+        "line_color": "#000000", "line_dash": "dotted"}, doc="""
+        Keyword arguments passed to ax.axvline in order to customize
+        the appearance of vertical lines indicating poles.""")
+    grid_line_kw = param.Dict(default={
+        "line_color": "#aaa", "line_dash": "dotted"}, doc="""
+        Keyword arguments used to customize the major grid lines.""")
     _fig = param.Parameter(default=None, doc="""
         The figure in which symbolic expressions will be plotted into.""")
 
@@ -182,11 +183,30 @@ class BokehBackend(Plot):
         if self._fig is None:
             self._fig = self.bokeh.plotting.figure(**kw)
         self._fig.axis.visible = self.axis
-        self._fig.grid.visible = self.grid
-        if self.minor_grid:
-            self._fig.grid.minor_grid_line_alpha = cfg["bokeh"]["minor_grid_line_alpha"]
-            self._fig.grid.minor_grid_line_color = self._fig.grid.grid_line_color[0]
-            self._fig.grid.minor_grid_line_dash = cfg["bokeh"]["minor_grid_line_dash"]
+
+        show_major_grid = True if self.grid else False
+        show_minor_grid = True if self.minor_grid else False
+        grid_lines_kw = {}
+        major_grid_line_kw = {}
+        minor_grid_line_kw = {
+            "minor_grid_line_alpha": cfg["bokeh"]["minor_grid_line_alpha"],
+            "minor_grid_line_color": self._fig.grid.grid_line_color[0],
+            "minor_grid_line_dash": cfg["bokeh"]["minor_grid_line_dash"]
+        }
+        if isinstance(self.grid, dict):
+            grid_lines_kw = self.grid
+        if show_minor_grid:
+            grid_lines_kw["minor_grid_line_alpha"] = cfg["bokeh"]["minor_grid_line_alpha"]
+            grid_lines_kw["minor_grid_line_color"] = self._fig.grid.grid_line_color[0]
+            grid_lines_kw["minor_grid_line_dash"] = cfg["bokeh"]["minor_grid_line_dash"]
+        if isinstance(self.minor_grid, dict):
+            grid_lines_kw = self.merge(
+                {}, grid_lines_kw, self.minor_grid)
+
+        self._fig.grid.visible = show_major_grid
+        for k, v in grid_lines_kw.items():
+            setattr(self._fig.grid, k, v)
+
         if self.invert_x_axis:
             self._fig.x_range.flipped = True